backend/src/query_generator/requirements.txt
anthropic==0.39.0
psycopg2-binary==2.9.9
httpx==0.27.2
httpcore==1.0.6
anyio==4.6.2.post1
sniffio==1.3.1
h11==0.14.0
certifi==2024.8.30
